Login logout cuoi cung cung dc (cmn >"<)

+++ Source/MySweetHome/MySweetHome-war/src/java/com/mysweethome/jsfmanagedbean/LoginBean.java
@@ -5,6 +5,7 @@
 package com.mysweethome.jsfmanagedbean;
 
 import com.mysweethome.entity.Member1;
+import com.mysweethome.helper.messages;
 import com.mysweethome.helper.operationSession;
 import com.mysweethome.session.Member1Facade;
 import java.io.IOException;
@@ -21,14 +22,15 @@ import org.primefaces.context.RequestContext;
  *
  * @author LongVu
  */
-@ManagedBean(name="LoginBean")
+@ManagedBean(name = "loginBean")
 @SessionScoped
 public class LoginBean {
+
     @EJB
-    private Member1Facade member1Facade;
-    private Member1 member=new Member1();
-    private String username;
-    private String password;
+    public Member1Facade member1Facade;
+    public Member1 member = new Member1();
+    public String username;
+    public String password;
 
     public Member1 getMember() {
         return member;
@@ -53,54 +55,64 @@ public class LoginBean {
     public void setPassword(String password) {
         this.password = password;
     }
+
     /**
      * Creates a new instance of LoginBean
      */
     public LoginBean() {
-        member= new Member1();
-        member1Facade= new Member1Facade();
+        member = new Member1();
+        member1Facade = new Member1Facade();
     }
-    
-    public void doLogin(ActionEvent actionEvent) {
+
+    public void doLogin() throws IOException {
         RequestContext context = RequestContext.getCurrentInstance();
-        FacesMessage msg = null;
         boolean loggedIn = false;
-        
+
         if (username.length() == 0) {
-            msg = new FacesMessage(FacesMessage.SEVERITY_WARN, "Login Error", "Username is not empty .");
+            messages.taoTB(FacesMessage.SEVERITY_WARN, "Login Error", "Username is not empty .");
         } else if (password.length() == 0) {
-            msg = new FacesMessage(FacesMessage.SEVERITY_WARN, "Login Error", "Password is not empty .");
+            messages.taoTB(FacesMessage.SEVERITY_WARN, "Login Error", "Password is not empty .");
         } else {
             Member1 mem = member1Facade.getUserName(username);
-            if (mem != null) {    
-                if(mem.getIsEnabled().equalsIgnoreCase("true")){
+            if (mem != null) {
+                if (mem.getIsEnabled().equalsIgnoreCase("true")) {
                     if (mem.getPassword().equals(password)) {
-                        loggedIn=true;
+                        loggedIn = true;
                         operationSession.createSession("username", mem.getUserName());
                         operationSession.createSession("role", mem.getRole());
-                        msg = new FacesMessage(FacesMessage.SEVERITY_INFO, "Welcome", mem.getUserName());                         
+                        messages.taoTB(FacesMessage.SEVERITY_INFO, "Welcome", mem.getUserName());
+
+//                        FacesContext context1=FacesContext.getCurrentInstance();
+//                        String viewid=context1.getViewRoot().getViewId();
+//                        ViewHandler handler=context1.getApplication().getViewHandler();
+//                        UIViewRoot root=handler.createView(context1, viewid);
+//                        root.setViewId(viewid);
+//                        context1.setViewRoot(root);
+                        FacesContext.getCurrentInstance().getExternalContext().redirect(FacesContext.getCurrentInstance().getExternalContext().getRequestContextPath() + "/index.jsf");
                     } else {
-                        msg= new FacesMessage(FacesMessage.SEVERITY_ERROR, "Password incorect", "Please input password");
+                        messages.taoTB(FacesMessage.SEVERITY_ERROR, "Password incorect", "Please input password");
                     }
-                }
-                else{
-                    msg=new FacesMessage(FacesMessage.SEVERITY_ERROR,"Account hasn't actived jet.","Please active your account before.");
+                } else {
+                    messages.taoTB(FacesMessage.SEVERITY_ERROR, "Account hasn't actived jet.", "Please active your account before.");
                 }
             } else {
-                msg=new FacesMessage(FacesMessage.SEVERITY_ERROR, "Username incorect", "Please input correct user name");
+                messages.taoTB(FacesMessage.SEVERITY_ERROR, "Username incorect", "Please input correct user name.");
             }
         }
-        
-        FacesContext.getCurrentInstance().addMessage(null, msg);
-        context.addCallbackParam("loggedIn", loggedIn);
     }
-    
-    public void doLogout() throws IOException{
-        operationSession.deleteGTsession("username");
-        operationSession.deleteGTsession("role");
-        this.username=null;
-        this.password=null;
+
+    public void back() throws IOException {
+        FacesContext.getCurrentInstance().getExternalContext().redirect(FacesContext.getCurrentInstance().getExternalContext().getRequestContextPath() + "/index.jsf");
+    }
+
+    public void doLogout() throws IOException {
+        if(operationSession.getSession("username")!=null)
+            operationSession.deleteGTsession("username");
+        if(operationSession.getSession("role")!=null)
+            operationSession.deleteGTsession("role");
+        this.username = null;
+        this.password = null;
         HttpServletResponse response = (HttpServletResponse) FacesContext.getCurrentInstance().getExternalContext().getResponse();
-        response.sendRedirect("../faces/index.xhtml");
+        response.sendRedirect("./index.jsf");
     }
 }
